media-source/SourceBufferList-SourceBuffer.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Test for SourceBufferList#SourceBuffer</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var FILE = 'test.webm';
var video = document.querySelector('video');
var mime = 'video/webm; codecs="vorbis,vp8"';

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
async_test(function(t) {
    if(!canPlayType.bind(t)(mime)) {
        return;
    }
    var mediaSource = new MediaSource();
    mediaSource.addEventListener('sourceopen', function(e){
        t.step(function() {
            assert_equals(mediaSource.sourceBuffers.SourceBuffer(0), undefined,
                          'If index is equal to the length attribute then return undefined.');
            assert_equals(mediaSource.sourceBuffers.SourceBuffer(1), undefined,
                          'If index is greater than the length attribute then return undefined.');

            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            assert_equals(mediaSource.sourceBuffers.SourceBuffer(0), sourceBuffer,
                          'Should return the index\'th SourceBuffer object in the list.');
            assert_equals(mediaSource.sourceBuffers.SourceBuffer(1), undefined,
                          'If index is equal to the length attribute then return undefined.');
            assert_equals(mediaSource.sourceBuffers.SourceBuffer(2), undefined,
                          'If index is greater than the length attribute then return undefined.');
        });
        t.done();
    });
    video.src = window.URL.createObjectURL(mediaSource);
},
'SourceBufferList#SourceBuffer : Check the behavior with the specified index parameter.');
</script>
</body>
</html>
